requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==1.10.9
google-generativeai==0.3.0
python-dotenv==1.0.0
